refactor and add application builder extensions to check marker service
