+++ src/Detection.Browser/src/Collections/InternetExplorer.cs
@@ -2,6 +2,7 @@
 // The Apache v2. See License.txt in the project root for license information.
 
 using System;
+using Wangkanai.Detection.Extensions;
 
 namespace Wangkanai.Detection.Collections
 {
